Refactored RequestHandler2s and made other core changes. 1. Refactored RequestHandler2s:    1. RequestHandler2s are now called ExecutionInterceptors and provide additional hooks into the request/response lifecycle.    2. Changed execution interceptor loading order to be more consisistent than RequestHandler2s.    3. Classpath-based interceptor loading now uses the current thread's classloader instead of the class's classloader.    4. Customer can now load service-specific interceptors from the classpath (this was previously limited functionality only available to the SDK).    5. Added new SdkRequest and SdkResponse base class on request/response objects.    6. Changed response object generation so that all responses will now be "OperationNameResponse", even if they're a simple wrapper for an inner object.    7. Changed signer and presigner interfaces to more closely match the response handler interface. 2. LegacyClientConfiguration has been entirely removed. The core SDK now uses ClientConfiguration:    1. Changed CRC32-from-compressed-data to be an SDK-internal advanced option on the ClientOverrideConfiguration in order to be consistent with all of the other client options.    2. Removed unused internal client constructors.    3. Split HttpClientDependencies into sync vs async variants.    4. HTTP request pipelines are now type safe in the sync vs async dependencies that they accept. 3. Removed legacy client builders. 4. Removed a large amount of the API gateway open SDK code. 5. Updated signing time offset to always be updated with each request to a service instead of just the first one. 6. Temporarily removed support for dry run requests in the EC2 client. 7. Migrated some of the shared logic in AsyncClientHandlerImpl and SyncClientHandlerImpl to a base class. They both still have a long way to go. 8. Changed ClientOverrideConfiguration to accept a v2 retry policy. 9. Made SdkHttpFullRequest and SdkHttpFullResponse more similar. They still have some inconsistencies to fix.

+++ core/src/main/java/software/amazon/awssdk/internal/http/timers/client/ClientExecutionTimer.java
@@ -50,7 +50,7 @@ public class ClientExecutionTimer implements AutoCloseable {
      * @return Implementation of {@link ClientExecutionAbortTrackerTaskImpl} to query the state of
      *         the task, provide it with up to date context, and cancel it if appropriate
      */
-    public ClientExecutionAbortTrackerTask startTimer(int clientExecutionTimeoutMillis) {
+    public ClientExecutionAbortTrackerTask startTimer(long clientExecutionTimeoutMillis) {
         if (isTimeoutDisabled(clientExecutionTimeoutMillis)) {
             return NoOpClientExecutionAbortTrackerTask.INSTANCE;
         } else if (executor == null) {
@@ -89,14 +89,14 @@ public class ClientExecutionTimer implements AutoCloseable {
         }
     }
 
-    private ClientExecutionAbortTrackerTask scheduleTimerTask(int clientExecutionTimeoutMillis) {
+    private ClientExecutionAbortTrackerTask scheduleTimerTask(long clientExecutionTimeoutMillis) {
         ClientExecutionAbortTask timerTask = new ClientExecutionAbortTaskImpl(Thread.currentThread());
         ScheduledFuture<?> timerTaskFuture = executor.schedule(timerTask, clientExecutionTimeoutMillis,
                                                                TimeUnit.MILLISECONDS);
         return new ClientExecutionAbortTrackerTaskImpl(timerTask, timerTaskFuture);
     }
 
-    private boolean isTimeoutDisabled(int clientExecutionTimeoutMillis) {
+    private boolean isTimeoutDisabled(long clientExecutionTimeoutMillis) {
         return clientExecutionTimeoutMillis <= 0;
     }
